src/python/requirements.txt
backcall==0.2.0
boto3==1.14.30
botocore==1.17.30
certifi==2020.6.20
chardet==3.0.4
Cython==0.29.21
decorator==4.4.2
docutils==0.15.2
enum34==1.1.10
future==0.18.2
idna==2.10
ijson==3.1.post0
indra==1.17.0
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
lxml==4.5.2
MarkupSafe==1.1.1
mpmath==1.1.0
ndex2==2.0.1
networkx==2.4
numpy==1.19.1
objectpath==0.6.1
obonet==0.2.5
pandas==1.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protmapper==0.0.17
ptyprocess==0.6.0
Pygments==2.6.1
pyjnius==1.3.0
pyparsing==2.4.7
pysb==1.9.1
pysolr==3.9.0
python-dateutil==2.8.1
pytz==2020.1
rdflib==4.2.2
requests==2.24.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
scipy==1.5.2
six==1.15.0
sympy==1.6.1
tqdm==4.48.0
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
